requirements.txt
lifelines==0.27.1
matplotlib==3.5.3
networkx==2.8.5
numpy==1.23.2
pandas==1.4.3
rdkit-pypi==2022.3.5
scikit-learn==1.1.2
scikit-plot==0.3.7
tensorflow-macos==2.9.1
tensorflow-addons==0.17.1
tqdm==4.64.0